Refactor for current progress
